+++ app/src/test/java/protect/card_locker/ImportURITest.java
@@ -40,7 +40,7 @@ public class ImportURITest {
         // Generate card
         Date date = new Date();
 
-        db.insertLoyaltyCard("store", "note", date, new BigDecimal("100"), null, BarcodeFormat.UPC_A.toString(), LoyaltyCardDbIds.BARCODE_TYPE, Color.BLACK, 1);
+        db.insertLoyaltyCard("store", "note", date, new BigDecimal("100"), null, BarcodeFormat.UPC_E.toString(), BarcodeFormat.UPC_A.toString(), BarcodeFormat.QR_CODE, Color.BLACK, 1);
 
         // Get card
         LoyaltyCard card = db.getLoyaltyCard(1);
@@ -53,6 +53,7 @@ public class ImportURITest {
 
         // Compare everything
         assertEquals(card.barcodeType, parsedCard.barcodeType);
+        assertEquals(card.barcodeId, parsedCard.barcodeId);
         assertEquals(card.cardId, parsedCard.cardId);
         assertEquals(card.headerColor, parsedCard.headerColor);
         assertEquals(card.note, parsedCard.note);
@@ -68,7 +69,7 @@ public class ImportURITest {
     public void ensureNoCrashOnMissingHeaderFields() throws InvalidObjectException
     {
         // Generate card
-        db.insertLoyaltyCard("store", "note", null, new BigDecimal("10.00"), Currency.getInstance("EUR"), BarcodeFormat.UPC_A.toString(), LoyaltyCardDbIds.BARCODE_TYPE, null, 0);
+        db.insertLoyaltyCard("store", "note", null, new BigDecimal("10.00"), Currency.getInstance("EUR"), BarcodeFormat.UPC_A.toString(), null, BarcodeFormat.QR_CODE, null, 0);
 
         // Get card
         LoyaltyCard card = db.getLoyaltyCard(1);
@@ -81,6 +82,7 @@ public class ImportURITest {
 
         // Compare everything
         assertEquals(card.barcodeType, parsedCard.barcodeType);
+        assertEquals(card.barcodeId, parsedCard.barcodeId);
         assertEquals(card.cardId, parsedCard.cardId);
         assertEquals(card.note, parsedCard.note);
         assertEquals(card.expiry, parsedCard.expiry);
@@ -88,7 +90,6 @@ public class ImportURITest {
         assertEquals(card.balanceType, parsedCard.balanceType);
         assertEquals(card.store, parsedCard.store);
         assertNull(parsedCard.headerColor);
-        assertNull(parsedCard.headerTextColor);
     }
 
     @Test
@@ -125,7 +126,8 @@ public class ImportURITest {
         }
 
         // Compare everything
-        assertEquals("ITF", parsedCard.barcodeType);
+        assertEquals(BarcodeFormat.ITF, parsedCard.barcodeType);
+        assertEquals(null, parsedCard.barcodeId);
         assertEquals("12345", parsedCard.cardId);
         assertEquals("note", parsedCard.note);
         assertEquals("store", parsedCard.store);