+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+Changes:
+
+- Add option to set a separate barcode value from card ID
+
 ## v1.12 (2021-03-30)
 
 Changes: